# swagger.json
{
   "openapi": "3.1.0",
   "info":{
    "title": "Postaway App",
    "version": "1.0.0"
   },
   "servers":[{
    "url": "http://localhost:3200",
    "description": "Dev server"

   }],
   "components": {
    "securitySchemes": {
        "bearerAuth":{
            "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"

        }
        
    }
   },
   "paths":{
...
    "/api/comments/{postId}/new": {
        "post": {
            "summary": "Add a new comment",
            "operationId": "addComment",
            "tags": ["Comment"],
            "security": [
                {
                    "bearerAuth": []
                }
            ],

            "parameters": [
                {
                    "in": "path",
                    "name": "postId",
                    "schema": {
                        "type": "string"
                    },
                    "required": true
                }
            ],

            "requestBody": {
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded" : {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "content": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },

            "responses": {
                "201": {
                    "description": "comment added successfully to post",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "userId": {
                                        "type": "string"
                                    },
                                    "postId": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "/api/comments/{postId}/all": {
        "get": {
            "summary": "get all comments on a given post",
            "operationId": "getComments",
            "tags": ["Comment"],
            "security": [
                {
                    "bearerAuth": []
                }
            ],

            "parameters": [
                {
                    "in": "path",
                    "name": "postId",
                    "schema": {
                        "type": "string"
                    }
                }
            ],

            "responses": {
                "200": {
                    "description": "array of all comments",
                    "content": {
                        "application/json": {
                            "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "userId": {
                                        "type": "string"
                                    },
                                    "postId": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    
                                    }
                                }
                            }
                        
                        }
                    }
                }
            },
            "400": {
                "description": "no comments made on the post",
                "content": {
                    "text/plain": {
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }

    }
   },
   "/api/comments/{postId}/{commentId}": {
    "post": {
        "summary": "update comment on a post",
        "operationId":"updateComment",
        "tags": ["Comment"],
        "security": [
            {
                "bearerAuth": []
            }
        ],
        "parameters": [
            {
                "in": "path",
                "required": true,
                "name": "postId",
                "schema": {
                    "type": "string"
                }
            },

            {
                "in": "path",
                "required": true,
                "name": "commentId",
                "schema": {
                    "type": "string"
                }
            }
        ],

        "requestBody": {
            "required": true,
            "content": {
                "application/x-www-form-urlencoded": {
                    "schema": {
                        "type": "object",
                        "properties": {
                            "content": {
                                "type": "string"
                            }
                        }
                    }
                }
            }        
        },

        "responses": {
            "201": {
                "description": "comment updated successfully",
                "content" :{
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "userId": {
                                        "type": "string"
                                    },
                                    "postId": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    
                                    }
                                }
                        }
                    }
                }
            },

            "400": {
                "description": "no comment found to update",
                "content": {
                    "text/plain": {
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },

    "delete": {
        "summary": "delete a comment",
        "operationId": "deleteComment",
        "tags": ["Comment"],
        "security": [
            {
                "bearerAuth": []
            }
        ],

        "parameters": [
            {
                "in": "path",
                "name": "postId",
                "schema": {
                    "type": "string"
                }
            },

            {
                "in": "path",
                "name": "commentId",
                "schema": {
                    "type": "string"
                }
            }
        ],

        "responses": {
            "201": {
                "description": "comment deleted successfully",
                "content": {
                    "text/plain": {
                        "schema": {
                            "type":"string"
                        }
                    }
                }
            },

            "400": {
                "description": "no comment found to delete",
                "content": {
                    "text/plain": {
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
   }
}
}